Avance EJB Tipo Cliente

+++ Pruebas/src/main/java/co/edu/uniquindio/ingesis/suturno/entidades/TipoClienteTest.java
@@ -70,7 +70,7 @@ public class TipoClienteTest {
 
 		TipoCliente tCliente = new TipoCliente();
 
-		tCliente.setId(3);
+		tCliente.setId((long) 3);
 		tCliente.setCodigo("gold");
 		tCliente.setNombre("Cliente Premium");
 		tCliente.setPrioridad(true);
